+++ obligatorio2-compgraf/Color.h
@@ -3,7 +3,7 @@
 #include <glm/glm.hpp>
 
 struct Color {
-	glm::vec3 rgb;
+	glm::dvec3 rgb;
 	int reflection;
 	int transmission;
 };
