+++ obligatorio2-compgraf/PositionLight.h
@@ -7,9 +7,9 @@ class PositionLight :
     public Light
 {
 public:
-    PositionLight(glm::vec3 position, Color color, float intensity);
-    glm::vec3 getPosition(); 
+    PositionLight(glm::dvec3 position, Color color, float intensity);
+    glm::dvec3 getPosition(); 
 private:
-    glm::vec3 position;
+    glm::dvec3 position;
 };
