+++ obligatorio2-compgraf/SceneObject.h
@@ -7,33 +7,33 @@ class Ray;
 class SceneObject;
 
 struct CollisionPoint {
-	glm::vec3 normal;
-	glm::vec3 position;
-	glm::vec3 hitDir;
-	float angle;
+	glm::dvec3 normal;
+	glm::dvec3 position;
+	glm::dvec3 hitDir;
+	double angle;
 	SceneObject* object;
-	float distanceFromOrigin;
+	double distanceFromOrigin;
 };
 
 class SceneObject
 {
 protected:
-	glm::vec3 position;
+	glm::dvec3 position;
 	
 public:
 	Color ambient;
 	Color specular;
 	Color diffuse;
-	float reflectionCoefficient;
-	float transmissionCoefficient;
-	float shininess;
-	float alpha;
+	double reflectionCoefficient;
+	double transmissionCoefficient;
+	double shininess;
+	double alpha;
 
 	bool solveQuadratic(const double& a, const double& b, const double& c, double& x0, double& x1);
 	virtual CollisionPoint* intersects(const Ray& ray);
-	float getReflectionCoefficient();
-	float getShininess();
-	float getTransmissionCoefficient();
-	float getMedium();
+	double getReflectionCoefficient();
+	double getShininess();
+	double getTransmissionCoefficient();
+	double getMedium();
 };
